requirements.txt
pyobjc
websockets
python-dotenv
